# docs/openapi/openapi-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Sample Restaurant API",
    "description": "Esta API demonstra endpoints de restaurante do sample.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/Mesas/{id}": {
      "get": {
        "tags": [
          "Mesas"
        ],
        "summary": "Método responsavel pela obtenção da Mesa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de identificação do objeto a ser pesquisado",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "A chamada precisa ser efetuada por um usuario autenticado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "O usuário esta autenticado, mas o não possui permissão para executar essa ação.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Retorna o objeto referente a ID informada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MesaViewModel"
                }
              }
            }
          },
          "404": {
            "description": "Caso o objeto não seja encontrada pela ID retorna null",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Requisicao invalida.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "Mesas"
        ],
        "summary": "Atualiza o Mesa no sistema.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de identificação do Mesa a ser atualiado",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MesaViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MesaViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MesaViewModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "A chamada precisa ser efetuada por um usuario autenticado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "O usuário esta autenticado, mas o não possui permissão para executar essa ação.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "Objeto atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResult"
                }
              }
            }
          },
          "400": {
            "description": "Não foi possivel executar a ação solicitada",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Mesas"
        ],
        "summary": "Exclui o Mesa do sistema.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de identificação do Mesa a ser atualiado",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "A chamada precisa ser efetuada por um usuario autenticado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "O usuário esta autenticado, mas o não possui permissão para executar essa ação.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "Objeto excluido com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomResult"
                }
              }
            }
          },
          "400": {
            "description": "Não foi possivel executar a ação solicitada",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "O objeto não foi encontrado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CustomResult": {
        "required": [
          "success",
          "data"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "description": "Objeto de retorno da chamada"
          }
        }
      },
      "ELocalizacaoMesa": {
        "type": "integer"
      },
...
      "MesaViewModel": {
        "required": [
          "numero",
          "localizacaoMesa"
        ],
        "type": "object",
        "properties": {
          "numero": {
            "maxLength": 50,
            "minLength": 1,
            "type": "string",
            "nullable": true
          },
          "lugares": {
            "maximum": 2147483647,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32"
          },
          "ativo": {
            "type": "boolean"
          },
          "localizacaoMesa": {
            "$ref": "#/components/schemas/ELocalizacaoMesa"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Informe somente o token JWT. A UI envia o prefixo Bearer no header Authorization.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}